ENT-5134 Add Postgres SessionFactoryFactory
