+++ node-api/src/main/resources/META-INF/services/net.corda.nodeapi.internal.persistence.CordaSessionFactoryFactory
@@ -1,1 +1,2 @@
 net.corda.nodeapi.internal.persistence.H2SessionFactoryFactory
+net.corda.nodeapi.internal.persistence.PostgresSessionFactoryFactory
